requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.0
mysql-connector==2.2.9
Pillow==7.1.2
PyMySQL==0.9.3
python-docx==0.8.10
requests==2.23.0
selenium==3.141.0
soupsieve==2.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
you-get==0.4.1432
